refactor: consolidate release preparation helper
